requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1

# LangChain + providers
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0

# Vector store
faiss-cpu>=1.8.0

# Web loading / parsing
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
